requirements.txt
Flask
flask_sqlalchemy
flask_talisman
flask_login
flask_wtf
WTForms
python-dotenv
bcrypt
cryptography
pyotp
mysql-connector
DateTime
email_validator
feedparser
flake8
bs4
requests
flask-talisman
